+++ lib/shopify_api/jwt_session_token.ex
@@ -50,21 +50,13 @@ defmodule ShopifyAPI.JWTSessionToken do
   def get_offline_token(%JOSE.JWT{} = jwt, token) do
     with {:ok, myshopify_domain} <- myshopify_domain(jwt),
          {:ok, app} <- app(jwt) do
-      case ShopifyAPI.AuthTokenServer.get(myshopify_domain, app.name) do
+      case ShopifyAPI.AuthTokenServer.get(myshopify_domain, app.handle) do
         {:ok, _} = resp ->
           resp
 
         {:error, _} ->
           Logger.warning("No token found, exchanging for new")
-
-          case ShopifyAPI.AuthRequest.request_offline_access_token(app, myshopify_domain, token) do
-            {:ok, token} ->
-              fire_post_login_hook(token)
-              {:ok, token}
-
-            error ->
-              error
-          end
+          request_offline_token(app, myshopify_domain, token)
       end
     else
       error ->
@@ -81,26 +73,40 @@ defmodule ShopifyAPI.JWTSessionToken do
     with {:ok, myshopify_domain} <- myshopify_domain(jwt),
          {:ok, app} <- app(jwt),
          {:ok, user_id} <- user_id(jwt) do
-      case ShopifyAPI.UserTokenServer.get_valid(myshopify_domain, app.name, user_id) do
+      case ShopifyAPI.UserTokenServer.get_valid(myshopify_domain, app.handle, user_id) do
         {:ok, _} = resp ->
           resp
 
         {:error, :invalid_user_token} ->
           Logger.debug("Expired or no user token found, exchanging for new")
-
-          case ShopifyAPI.AuthRequest.request_online_access_token(app, myshopify_domain, token) do
-            {:ok, user_token} ->
-              fire_post_login_hook(user_token)
-              {:ok, user_token}
-
-            error ->
-              error
-          end
+          request_online_token(app, myshopify_domain, token)
       end
     else
       error ->
         Logger.warning("failed getting required informatio from the JWT #{inspect(error)}")
         {:error, :invalid_session_token}
+    end
+  end
+
+  defp request_offline_token(app, myshopify_domain, token) do
+    case ShopifyAPI.AuthRequest.request_offline_access_token(app, myshopify_domain, token) do
+      {:ok, token} ->
+        fire_post_login_hook(token)
+        {:ok, token}
+
+      error ->
+        error
+    end
+  end
+
+  defp request_online_token(app, myshopify_domain, token) do
+    case ShopifyAPI.AuthRequest.request_online_access_token(app, myshopify_domain, token) do
+      {:ok, token} ->
+        fire_post_login_hook(token)
+        {:ok, token}
+
+      error ->
+        error
     end
   end