+++ CHANGELOG.md
@@ -4,13 +4,15 @@
   If the REST client is used you must include the new REST client repo:
   `{:shopify_api_rest, github: "orbit-apps/elixir-shopifyapi-rest", tag: "v0.1.0"}`
 - BREAKING: AppServer now defaults to a single app instance, this is a compile env if you want to use the old multi app config add `config :shopify_api, :app_server, :multi_app` to your `config/config.exs`
+- BREAKING: App.handle is used everywhere instead of App.name for keys on AuthToken, UserAuthToken, and Shop
+- BREAKING: The Shop struct now has a `myshopify_domain` field instead of `domain`
+- BREAKING: old Plugs.Webhook has been removed
 - New: Single app mode for AppServer, is API compatible with the multi app setup. This greatly simplifies the most common setup, one app <> one phoenix setup.
 - New: Add handle and raw app config to the App struct
 - New: App.new/1 function to load app from parsed Shopify app config toml file
 - New: GraphQL requests through [Req](https://hexdocs.pm/req/Req.html) are now done with GraphQLQuery modules and return GraphQLResponses. Ideally we will deprecate the previoud GraphQL method once people have had a chance to move over from the old method.
 - New: Add Scopes context and Scope protocol. Change GraphQL queries to expect scopes. AuthToken can be used as a scope as a fallback via the defimpl in the AuthToken file.
 - New: Reworked webhook flow, [check readme](README.md#Webhooks) for details on how to use
-- Deprecation: old Plugs.Webhook is being replaced and will be removed eventually
 
 ## 0.16.6